basic-proto: correct the type of local variables
raw_stream_pos, raw_buffer_size, buffer_pos and raw_buffer_leftover_size
variables are local copies of the LogProtoBufferedServerState's members.

This commit corrects these local declarations to match the original type.

+++ modules/basic-proto/basic-proto.c
@@ -1042,7 +1042,7 @@ log_proto_buffered_server_apply_state(LogProto *s, StateHandler *state_handler)
   LogProtoBufferedServerState *state;
   gint fd;
   gboolean new_run;
-  guint64 raw_stream_pos, raw_buffer_size, buffer_pos;
+  gint64 raw_stream_pos, raw_buffer_size, buffer_pos;
 
   fd = self->super.transport->fd;
   if (self->state_handler)
@@ -1646,7 +1646,7 @@ log_proto_buffered_server_fetch(LogProto *s, const guchar **msg, gsize *msg_len,
   LogProtoBufferedServerState *state = log_proto_buffered_server_get_state(self);
   LogProtoStatus result = self->status;
   LogProtoServerOptions *options = (LogProtoServerOptions *)self->super.options;
-  gint raw_buffer_leftover_size;
+  guint8 raw_buffer_leftover_size;
 
   if (G_UNLIKELY(!self->buffer))
     {
